Extract test driver logic into it's own class.
